requirements.txt
# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Financial data
yfinance>=0.2.0
TA-Lib>=0.4.0

# Visualization
mplfinance>=0.12.0
matplotlib>=3.7.0
plotly>=5.0.0

# Backtesting
backtrader>=1.9.0

# Configuration
pyyaml>=6.0

# Utilities
requests>=2.28.0
python-dotenv>=1.0.0
